refactor: change the syntax for asserting valid states
The Lexer class now has a method returning a boolean that tests if the
internal curent token has one of the token ids passed as argument.
We also have a method that throws an exception containing information
about the current token being processed, and the position in the source
code where we're lexing right now.
Those two methods can be used in conjunction to easily handle syntax
errors spotted in the source.